Add useNumberFieldInput hook
