+++ packages/react/src/number-field/decrement/NumberFieldDecrement.test.tsx
@@ -10,24 +10,31 @@ import { NumberFieldRootContext } from '../root/NumberFieldRootContext';
 
 const testContext = {
   allowInputSyncRef: { current: false },
+  autoFocus: false,
   disabled: false,
   formatOptionsRef: { current: undefined },
-  getInputProps: (externalProps) => externalProps,
+  getAllowedNonNumericKeys: () => [],
   getScrubAreaProps: (externalProps) => externalProps,
   getScrubAreaCursorProps: (externalProps) => externalProps,
   getStepAmount: NOOP,
   id: 'id',
   incrementValue: NOOP,
+  inputMode: 'numeric',
   inputRef: { current: null },
   inputValue: '',
   intentionalTouchCheckTimeoutRef: { current: -1 },
+  invalid: false,
   isPressedRef: { current: false },
   isScrubbing: false,
+  max: undefined,
   maxWithDefault: 100,
   mergedRef: (_node) => {},
+  min: undefined,
   minWithDefault: 0,
+  name: 'NumberField',
   movesAfterTouchRef: { current: 0 },
   readOnly: false,
+  required: false,
   scrubAreaRef: { current: null },
   scrubAreaCursorRef: { current: null },
   scrubHandleRef: {
@@ -37,6 +44,7 @@ const testContext = {
       teleportDistance: 0,
     },
   },
+  setInputValue: NOOP,
   setValue: NOOP,
   state: {
     value: null,
