+++ packages/react/src/number-field/group/NumberFieldGroup.test.tsx
@@ -8,24 +8,31 @@ import { NOOP } from '../../utils/noop';
 
 const testContext = {
   allowInputSyncRef: { current: false },
+  autoFocus: false,
   disabled: false,
   formatOptionsRef: { current: undefined },
-  getInputProps: (externalProps) => externalProps,
+  getAllowedNonNumericKeys: () => [],
   getScrubAreaProps: (externalProps) => externalProps,
   getScrubAreaCursorProps: (externalProps) => externalProps,
   getStepAmount: NOOP,
   id: 'id',
   incrementValue: NOOP,
+  inputMode: 'numeric',
   inputRef: { current: null },
   inputValue: '',
   intentionalTouchCheckTimeoutRef: { current: -1 },
+  invalid: false,
   isPressedRef: { current: false },
   isScrubbing: false,
+  max: undefined,
   maxWithDefault: 100,
   mergedRef: (_node) => {},
+  min: undefined,
   minWithDefault: 0,
+  name: 'NumberField',
   movesAfterTouchRef: { current: 0 },
   readOnly: false,
+  required: false,
   scrubAreaRef: { current: null },
   scrubAreaCursorRef: { current: null },
   scrubHandleRef: {
@@ -35,6 +42,7 @@ const testContext = {
       teleportDistance: 0,
     },
   },
+  setInputValue: NOOP,
   setValue: NOOP,
   state: {
     value: null,
