+++ packages/react/src/number-field/input/NumberFieldInput.test.tsx
@@ -4,16 +4,64 @@ import { act, screen, fireEvent } from '@mui/internal-test-utils';
 import { NumberField } from '@base-ui-components/react/number-field';
 import { createRenderer, describeConformance } from '#test-utils';
 import { NumberFieldRootContext } from '../root/NumberFieldRootContext';
+import { NOOP } from '../../utils/noop';
 
 const testContext = {
-  getInputProps: (externalProps) => externalProps,
+  allowInputSyncRef: { current: false },
+  autoFocus: false,
+  disabled: false,
+  formatOptionsRef: { current: undefined },
+  getAllowedNonNumericKeys: () => [],
+  getScrubAreaProps: (externalProps) => externalProps,
+  getScrubAreaCursorProps: (externalProps) => externalProps,
+  getStepAmount: NOOP,
+  id: 'id',
+  incrementValue: NOOP,
+  inputMode: 'numeric',
+  inputRef: { current: null },
+  inputValue: '',
+  intentionalTouchCheckTimeoutRef: { current: -1 },
+  invalid: false,
+  isPressedRef: { current: false },
+  isScrubbing: false,
+  max: undefined,
+  maxWithDefault: 100,
+  mergedRef: (_node) => {},
+  min: undefined,
+  minWithDefault: 0,
+  name: 'NumberField',
+  movesAfterTouchRef: { current: 0 },
+  readOnly: false,
+  required: false,
+  scrubAreaRef: { current: null },
+  scrubAreaCursorRef: { current: null },
+  scrubHandleRef: {
+    current: {
+      direction: 'horizontal',
+      pixelSensitivity: 0,
+      teleportDistance: 0,
+    },
+  },
+  setInputValue: NOOP,
+  setValue: NOOP,
   state: {
     value: null,
     required: false,
     disabled: false,
     invalid: false,
     readOnly: false,
+    scrubbing: false,
+    touched: false,
+    dirty: false,
+    inputValue: '',
+    valid: true,
+    filled: false,
+    focused: false,
   },
+  startAutoChange: NOOP,
+  stopAutoChange: NOOP,
+  value: null,
+  valueRef: { current: null },
 } as NumberFieldRootContext;
 
 describe('<NumberField.Input />', () => {
